* fixed lots of warnings about casting   should fix compilation on iphone as well
git-svn-id: svn://svn.camaya.net/gloox/branches/1.0@4135 bb60cc1b-8fe5-0310-9879-aca5d778a633

+++ src/connectionhttpproxy.cpp
@@ -115,7 +115,7 @@ namespace gloox
       m_connection->cleanup();
   }
 
-  void ConnectionHTTPProxy::getStatistics( int& totalIn, int& totalOut )
+  void ConnectionHTTPProxy::getStatistics( long int& totalIn, long int& totalOut )
   {
     if( m_connection )
       m_connection->getStatistics( totalIn, totalOut );
